requirements.txt
mbtest==2.12.0
pytest==8.1.1
python-dotenv==1.0.1